backend/requirements.txt
fastapi[all]
uvicorn
httpx
python-dotenv
polyline
starlette
aiofiles
python-multipart
google-genai
psycopg2-binary
requests
tensorflow==2.17 
opencv-python 
numpy